Added missing libs.deployDir
Debug deploy target appeared to be missing libs.deployDir , which was preventing a successful debug deploy from eclipse. I"m not sure if this is all of the required changes (as I still have to manually restart the remote java application), but this does allow the deploy to go through as expected.