Fix roboRIO image version detection in presence of EOL.
EOL usually does not exist in this data, but can be set via the Comments field
of the roboRIO system configuration.

On Windows, the post task erroneously turns UTF-16LE 0x0A00 (LF) in the
returned data into non-UTF16 0x0D0A00 (CRLF).  This causes the UTF16-LE
file read to fail.

Attempts to fix this and still write to a file failed due to similar issues
with other Ant tasks.  Instead, we now keep the data in a property and simply
remove all the nul characters.  While this will corrupt any actual UTF-16
characters, we use a regex to find the version information, which is always
ASCII.