Implement TCP netconsole.
For 2018, FRC_NetComm provides a TCP netconsole on port 1741.
This is a tagged data stream with both error/warning and console
information.

The console is now colorized similar to the DS console.
The console font and colors are customizable.
The console supports timestamps.

Currently rendering to text is done as messages arrive.  This means that
options such as enabling/disabling timestamping do not affect previously
received messages.

The NetComm robot implementation requires a DS connection for console output,
so this is noted in the startup text of the console.